+++ src/extension.ts
@@ -1,10 +1,11 @@
 import * as vscode from "vscode";
 import { Commands } from './enum';
-import { StatusbarUi } from "./StatusBarUI";
-import Server from "./server";
 import HomePage from './HomePage';
+import Server from "./server";
+import { StatusbarUi } from "./StatusBarUI";
 
 export function activate(context: vscode.ExtensionContext) {
+
   const server = new Server();
 
   // Generate Mock
@@ -14,7 +15,13 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(vscode.commands.registerCommand(Commands.START_SERVER, server.restartServer));
 
   // Stop Server
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.STOP_SERVER, server.stopServer));
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.STOP_SERVER, () => server.stopServer()));
+
+  // Reset Server
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.RESET_SERVER, () => server.resetServer()));
+
+  // Reset and Restart Server
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.RESET_AND_RESTART, (args) => server.restartServer(args, true)));
 
   // Switch Environment
   context.subscriptions.push(vscode.commands.registerCommand(Commands.SWITCH_ENVIRONMENT, server.switchEnvironment));
@@ -24,6 +31,9 @@ export function activate(context: vscode.ExtensionContext) {
 
   // Set Port
   context.subscriptions.push(vscode.commands.registerCommand(Commands.SET_PORT, server.setPort));
+
+  // Set Root
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.SET_ROOT, server.setRoot));
 
   // Create Sample Files
   context.subscriptions.push(vscode.commands.registerCommand(Commands.GENERATE_MOCK_FILES, server.generateMockFiles));