+++ src/server.ts
@@ -1,18 +1,50 @@
+import { MockServer } from "@r35007/mock-server";
+import { HAR } from '@r35007/mock-server/dist/server/types/common.types';
+import * as ValidTypes from "@r35007/mock-server/dist/server/types/valid.types";
+import { cleanDb, createSampleFiles, extractDbFromHAR } from '@r35007/mock-server/dist/server/utils';
+import { requireData } from '@r35007/mock-server/dist/server/utils/fetch';
+import * as fs from 'fs';
 import * as path from 'path';
 import { Commands } from './enum';
 import { Prompt } from './prompt';
 import { Settings } from './Settings';
 import { StatusbarUi } from './StatusBarUI';
 import { Utils } from './utils';
-import { Db } from '@r35007/mock-server/dist/server/model';
-import { createSampleFiles, getDbSnapShot } from '@r35007/mock-server/dist/server/utils';
 
-export default class MockServer extends Utils {
+export default class MockServerExt extends Utils {
+  mockServer!: MockServer;
+
   constructor() {
     super();
     this.output.appendLine('\nMock Server Server Initiated');
+    this.createServer();
     StatusbarUi.init();
   }
+
+  createServer = () => {
+    this.mockServer = new MockServer({ root: Settings.rootPath });
+  };
+
+  destroyServer = async () => {
+    await MockServer.Destroy(this.mockServer);
+  };
+
+  resetServer = async (isRestarting?: boolean) => {
+    this.output.appendLine(`\n[${new Date().toLocaleTimeString()}] [Running] Server Reset initiated`);
+    this.output.appendLine(`[${new Date().toLocaleTimeString()}] Server Resetting`);
+    try {
+      StatusbarUi.working('Resetting...');
+      await this.destroyServer();
+      this.createServer();
+      !isRestarting && StatusbarUi.startServer(150, () => Prompt.showPopupMessage('Server Reset Done', 'info'));
+      this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Server Reset Done`);
+    } catch (err: any) {
+      this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Server Failed to Reset`);
+      this.output.appendLine(err);
+      const statsMsg = `Server Failed to Reset. \n${err.message}`;
+      StatusbarUi.stopServer(0, Settings.port, () => Prompt.showPopupMessage(statsMsg, 'error'));
+    }
+  };
 
   transformToMockServerDB = async (args?: any) => {
     this.output.appendLine(`\n[${new Date().toLocaleTimeString()}] [Running] Data Transform initiated`);
@@ -25,13 +57,13 @@ export default class MockServer extends Utils {
           reverse: Settings.reverse,
           isSnapshot: true,
         };
-        const db = this.mockServer.getValidDb(
-          args?.fsPath || document?.uri?.fsPath,
-          [],
-          options,
-          Settings.entryCallback,
-          Settings.finalCallback
-        );
+        const userData = requireData(args?.fsPath || document?.uri?.fsPath) as HAR;
+        const db = extractDbFromHAR(
+          userData,
+          Settings._harEntryCallback,
+          Settings._harDbCallback
+        ) || userData;
+        cleanDb(db);
         this.writeFile(
           JSON.stringify(db, null, '\t'),
           fileName,
@@ -49,31 +81,43 @@ export default class MockServer extends Utils {
     }
   };
 
-  setPort = async (args?: any) => {
+  setPort = async (_args?: any) => {
     const port = await Prompt.showInputBox('Enter Port Number', Settings.port);
     if (port) {
       Settings.port = parseInt(port);
+      Prompt.showPopupMessage(`Port Number Set to ${port}`, 'info')
       this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Port Number Set to ${port}`);
     }
   };
 
-  startServer = async (txt: string, dbPath?: string) => {
+  setRoot = async (args?: any) => {
+    if (!args?.fsPath) { return; }
+    const stat = fs.statSync(args.fsPath);
+    const rootPath = stat.isFile() ? path.dirname(args.fsPath) : args.fsPath;
+    Settings.rootPath = rootPath;
+    Prompt.showPopupMessage(`Root Path Set to ${rootPath}`, 'info')
+    this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Root Path Set to ${rootPath}`);
+  };
+
+  startServer = async (txt: 'Start' | 'Restart', dbPath?: string) => {
     this.output.appendLine(`\n[${new Date().toLocaleTimeString()}] [Running] Server ${txt} initiated`);
     try {
-      const paths = Settings.paths;
-      const _dbPath = dbPath || paths.db;
       this.output.appendLine(`[${new Date().toLocaleTimeString()}] Server ${txt}ing...`);
       StatusbarUi.working(`${txt}ing...`);
 
-      const db = (await this.getDbWithEnv(_dbPath?.replace(/\\/g, '/'))) as Db;
+      const paths = Settings.paths;
+      const _dbPath = dbPath || paths.db;
+      const db = (await this.getDbWithEnv(_dbPath?.replace(/\\/g, '/'))) as ValidTypes.Db;
+
       this.mockServer.setConfig(Settings.config);
-      await this.mockServer.launchServer(db, paths.middleware, paths.injectors, paths.rewriters, paths.store);
+
+      await this.mockServer.launchServer(db, paths.injectors, paths.middleware, paths.rewriters, paths.store);
+
       this.restartOnChange(this.restartServer);
+
       const statusMsg = `Server is ${txt}ed at port : ${Settings.port}`;
       StatusbarUi.stopServer(150, Settings.port, () => Prompt.showPopupMessage(statusMsg, 'info'));
-      this.output.appendLine(
-        `[${new Date().toLocaleTimeString()}] [Done] Server is ${txt}ed at port : ${Settings.port}`
-      );
+      this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Server is ${txt}ed at port : ${Settings.port}`);
     } catch (err: any) {
       this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Server Failed to ${txt}`);
       this.output.appendLine(err);
@@ -82,15 +126,15 @@ export default class MockServer extends Utils {
     }
   };
 
-  stopServer = async () => {
+  stopServer = async (isRestarting?: boolean) => {
     this.output.appendLine(`\n[${new Date().toLocaleTimeString()}] [Running] Server Stop initiated`);
+    this.output.appendLine(`[${new Date().toLocaleTimeString()}] Server Stopping`);
     try {
-      this.output.appendLine(`[${new Date().toLocaleTimeString()}] Server Stopping`);
       if (this.mockServer.server) {
         StatusbarUi.working('Stopping...');
         await this.mockServer.stopServer();
         await this.stopWatchingChanges();
-        StatusbarUi.startServer(150, () => Prompt.showPopupMessage('Server is Stopped', 'info'));
+        !isRestarting && StatusbarUi.startServer(150, () => Prompt.showPopupMessage('Server is Stopped', 'info'));
         this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Server is Stopped`);
       } else {
         this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] No Server to Stop`);
@@ -105,21 +149,10 @@ export default class MockServer extends Utils {
     }
   };
 
-  restartServer = async (args?: any) => {
+  restartServer = async (args?: any, shouldResetServer?: boolean) => {
     if (this.mockServer.server) {
-      try {
-        this.output.appendLine(`\n[${new Date().toLocaleTimeString()}] [Running] Server Stop initiated`);
-        this.output.appendLine(`[${new Date().toLocaleTimeString()}] Server Stopping`);
-        StatusbarUi.working('Stopping...');
-        await this.mockServer.stopServer();
-        await this.stopWatchingChanges();
-        this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Server Stopped`);
-        await this.startServer('Restart', args?.fsPath);
-      } catch (err: any) {
-        this.output.appendLine(`[${new Date().toLocaleTimeString()}] [Done] Server Failed to Stop`);
-        StatusbarUi.stopServer(0, Settings.port);
-        this.output.appendLine(err);
-      }
+      shouldResetServer ? await this.resetServer(true) : await this.stopServer(true);
+      await this.startServer('Restart', args?.fsPath);
     } else {
       await this.startServer('Start', args?.fsPath);
     }
@@ -169,19 +202,18 @@ export default class MockServer extends Utils {
     }
   };
 
-  getDbSnapshot = async (args?: any) => {
+  getDbSnapshot = async (_args?: any) => {
     this.output.appendLine(`\n[${new Date().toLocaleTimeString()}] [Running] Db Snapshot initiated`);
     const writable = await this.getWritable(['.json'], Commands.TRANSFORM_TO_MOCK_SERVER_DB, true);
     if (writable) {
       this.output.appendLine(`[${new Date().toLocaleTimeString()}] Getting Db Snapshot..`);
       const { editor, document, textRange } = writable;
       try {
-        const db = JSON.parse(JSON.stringify(this.mockServer.data.db));
-        const snapShot = getDbSnapShot(db);
-        const snapShotPath =
-          args?.fsPath || path.join(Settings.paths.snapshotDir || 'snapshots', `/db-${Date.now()}.json`);
+        const db = JSON.parse(JSON.stringify(this.mockServer.db));
+        cleanDb(db);
+        const snapShotPath = path.join(Settings.paths.snapshotDir || 'snapshots', `/db-${Date.now()}.json`);
         this.writeFile(
-          JSON.stringify(snapShot, null, '\t'),
+          JSON.stringify(db, null, '\t'),
           snapShotPath,
           'Db Snapshot retrieved Successfully',
           editor,
@@ -199,7 +231,7 @@ export default class MockServer extends Utils {
 
   generateMockFiles = async (args?: any) => {
     this.output.appendLine('\nCreating Samples...');
-    createSampleFiles(args?.fsPath || Settings.paths.root);
+    createSampleFiles(args?.fsPath || Settings.rootPath);
     this.output.appendLine('Sample files created !');
   };
 }